requirements.txt
Django==3.2
Pillow
django-cleanup==6.0.0
django-filter==21.1
djangorestframework==3.13.1
django-cors-headers
django-modeltranslation==0.18.5
djangorestframework-simplejwt==4.0.0
django-ckeditor==6.3.2
django-phonenumber-field==6.0.0
django-mptt==0.14.0
drf-yasg==1.20.0
phonenumbers==8.12.57
PyJWT==1.7.1
psycopg2
psycopg2-binary
gunicorn
django-extra-fields